src/main/java/com/example/springbootboard/service/ArticleCommentService.java
package com.example.springbootboard.service;

import com.example.springbootboard.dto.ArticleCommentDto;
import com.example.springbootboard.repository.ArticleCommentRepository;
import com.example.springbootboard.repository.ArticleRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import java.util.List;
import java.awt.*;

@RequiredArgsConstructor
@Transactional
@Service
public class ArticleCommentService {

    private final ArticleRepository articleRepository;
    private final ArticleCommentRepository articleCommentRepository;

    @Transactional(readOnly = true)
    public List<ArticleCommentDto> searchArticleComment(Long articleId) {
        return List.of();
    }
}
